@@ -3,6 +3,7 @@
   "description": "GitHub Action to get last N tags from a repository",
   "version": "v1",
   "author": "hakwerk",
+  "type": "module",
   "private": true,
   "homepage": "https://github.com/hakwerk/gha-git-repo-tags",
   "repository": {
@@ -13,11 +14,10 @@
     "url": "https://github.com/hakwerk/gha-git-repo-tags/issues"
   },
   "keywords": [
-    "GitHub",
-    "Actions",
-    "JavaScript",
-    "Repository",
-    "Tags"
+    "actions",
+    "github",
+    "repository",
+    "tags"
   ],
   "exports": {
     ".": "./dist/index.js"
@@ -27,60 +27,37 @@
   },
   "scripts": {
     "bundle": "npm run format:write && npm run package",
-    "ci-test": "npx jest",
+    "ci-test": "NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 npx jest",
     "coverage": "npx make-coverage-badge --output-path ./badges/coverage.svg",
     "format:write": "npx prettier --write .",
     "format:check": "npx prettier --check .",
-    "lint": "npx eslint . -c ./.github/linters/.eslintrc.yml",
-    "package": "npx ncc build src/index.js -o dist --source-map --license licenses.txt",
+    "lint": "npx eslint .",
+    "local-action": "npx local-action . src/main.js .env",
+    "package": "npx rollup --config rollup.config.js",
     "package:watch": "npm run package -- --watch",
-    "test": "npx jest",
+    "test": "NODE_OPTIONS=--experimental-vm-modules NODE_NO_WARNINGS=1 npx jest",
     "all": "npm run format:write && npm run lint && npm run test && npm run coverage && npm run package"
   },
   "license": "MIT",
-  "eslintConfig": {
-    "extends": "./.github/linters/.eslintrc.yml"
-  },
-  "jest": {
-    "verbose": true,
-    "clearMocks": true,
-    "testEnvironment": "node",
-    "moduleFileExtensions": [
-      "js"
-    ],
-    "testMatch": [
-      "**/*.test.js"
-    ],
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "/dist/"
-    ],
-    "coverageReporters": [
-      "json-summary",
-      "text",
-      "lcov"
-    ],
-    "collectCoverage": true,
-    "collectCoverageFrom": [
-      "./src/**"
-    ]
-  },
   "dependencies": {
     "@actions/core": "^1.11.1",
     "@actions/github": "^6.0.0",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "@babel/core": "^7.26.0",
-    "@babel/eslint-parser": "^7.25.9",
-    "@babel/preset-env": "^7.26.0",
-    "@vercel/ncc": "^0.38.3",
-    "babel-preset-jest": "^29.6.3",
-    "eslint": "^8.57.0 <10.0.0",
-    "eslint-plugin-github": "^5.1.4",
+    "@eslint/compat": "^1.2.3",
+    "@jest/globals": "^29.7.0",
+    "@rollup/plugin-commonjs": "^28.0.1",
+    "@rollup/plugin-node-resolve": "^15.3.0",
+    "eslint": "^9.18.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jest": "^28.10.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "jest": "^29.7.0",
     "make-coverage-badge": "^1.2.0",
-    "prettier": "^3.4.2"
+    "prettier": "^3.4.2",
+    "prettier-eslint": "^16.3.0",
+    "rollup": "^4.27.1"
   }
 }
