@@ -1,6 +1,8 @@
 /**
- * The entrypoint for the action.
+ * The entrypoint for the action. This file simply imports and runs the action's
+ * main logic.
  */
-const { run } = require('./main')
+import { run } from './main.js'
 
+/* istanbul ignore next */
 run()
